Add Comments

+++ ImageBase.h
@@ -1,7 +1,3 @@
-//
-// Created by niccoboa on 20/02/2024.
-//
-
 #ifndef KERNELPROCESSING_IMAGEBASE_H
 #define KERNELPROCESSING_IMAGEBASE_H
 
@@ -13,14 +9,17 @@ public:
     virtual int getHeight() const = 0;
     virtual void setWidth(int w) = 0;
     virtual void setHeight(int h) = 0;
-    virtual int getPixel(int x, int y, int c) const = 0;
-    virtual void setPixel(int x, int y, int c, int value) = 0;
+    virtual float getPixel(int x, int y, int c) const = 0;
+    virtual void setPixel(int x, int y, int c, float value) = 0;
     // virtual const std::vector<std::vector<unsigned char>> & getData() const = 0;
     virtual void setData(const std::vector<std::vector<unsigned char>>& newData) = 0;
     virtual int operator()(int x, int y) const = 0;
 
     virtual Channel getChannels() const = 0;
     virtual ~ImageBase() = default;
+
+    // virtual ImageBase* clone() const = 0;
+
 };
